refactor: extract logout to its own module
