+++ src/cargo/ops/registry/mod.rs
@@ -3,6 +3,7 @@
 //! [1]: https://doc.rust-lang.org/nightly/cargo/reference/registry-web-api.html
 
 mod login;
+mod logout;
 mod publish;
 mod search;
 
@@ -29,6 +30,7 @@ use crate::util::IntoUrl;
 use crate::{drop_print, drop_println, version};
 
 pub use self::login::registry_login;
+pub use self::logout::registry_logout;
 pub use self::publish::publish;
 pub use self::publish::PublishOpts;
 pub use self::search::search;
@@ -348,42 +350,6 @@ impl HttpTimeout {
         handle.low_speed_limit(self.low_speed_limit)?;
         Ok(())
     }
-}
-
-pub fn registry_logout(config: &Config, reg: Option<&str>) -> CargoResult<()> {
-    let source_ids = get_source_id(config, None, reg)?;
-    let reg_cfg = auth::registry_credential_config(config, &source_ids.original)?;
-    let reg_name = source_ids.original.display_registry_name();
-    if reg_cfg.is_none() {
-        config.shell().status(
-            "Logout",
-            format!("not currently logged in to `{}`", reg_name),
-        )?;
-        return Ok(());
-    }
-    auth::logout(config, &source_ids.original)?;
-    config.shell().status(
-        "Logout",
-        format!(
-            "token for `{}` has been removed from local storage",
-            reg_name
-        ),
-    )?;
-    let location = if source_ids.original.is_crates_io() {
-        "<https://crates.io/me>".to_string()
-    } else {
-        // The URL for the source requires network access to load the config.
-        // That could be a fairly heavy operation to perform just to provide a
-        // help message, so for now this just provides some generic text.
-        // Perhaps in the future this could have an API to fetch the config if
-        // it is cached, but avoid network access otherwise?
-        format!("the `{reg_name}` website")
-    };
-    config.shell().note(format!(
-        "This does not revoke the token on the registry server.\n    \
-        If you need to revoke the token, visit {location} and follow the instructions there."
-    ))?;
-    Ok(())
 }
 
 pub struct OwnersOptions {
